Picture (BMP/JPEG) Manipulation Programs
Recover is capable of taking a stream of continuous bits from a 'wiped' memory storage device that has lost the specific locations to a series of JPEG files. It iterates over the bits and outputs JPEG files for each complete picture it finds until the end of the memory.
Resize takes an input BMP file and outputs a new BMP file that is scaled by an input parameter factor.
Whodunit decodes an encoded BMP file who's pixels have been manipulated to hide a message.